Specify sleep or not for all internal calls to getCurrentActivity
Many of the methods calling getCurrentActivity() don't rely on the
delay that is introduced by the no-argument method. Often, the method is
used to get the Activity only to fetch some other child object, and the 
delays that build up will slow down tests unnecessarily. 

For the case where the delay is relied on, this is made explicit.
Everywhere else, we make it explicit that we don't want the delay.

The alternative, of changing the default behaviour of the no-argument
constructor to return without delay would change the behaviour of the
associated Solo method, breaking compatibility with other's existing 
tests.