Move sleep() calls from start of wait loops to end.
When waiting for a condition to be true, it's better to check the
condition first, and then sleep if it isn't true before checking again.
This speeds up the (fairly common?) case when the condition is already
true before we start waiting for it to become true. 

This change re-orders the sleeps around so we always consistently check
the condition before sleeping.