Remove the sleep that always precedes the fetching of views
The only time that this callback is called is at the start of a 'wait
for timeout' loop in searchFor(Callable, String, int, etc...). As such,
if the first thing that happens is a sleep(), the search is always
delayed by the duration of that sleep, even if the search target is
already present. 

As the 'wait for timeout' loop in question already has a sleep in an 
appropriate place, the sleep in the callback is unnecessary.