Remove sleep from clickOnScreen
This is the only clickOn... method that has an explicit sleep before
performing the click, and there's no mention in the Javadoc to indicate
that we're intentionally sleeping here. It's not clear what the benefit
of sleeping here is, and it unexpectedly slows down tests for callers.
Removing it to speed up tests.