Update Scroller.java
I was running some tests but in my case scrolling to the side only by 0.55 the screen's width did't change my view totally, so I thought about scrolltoSideAllTheWay which takes all the screen width 0.99 and scrolls. Of course this could be done in a more intelligent way but since I am in a hurry I took the easiest Path.