Fix: issue-579
If a screenshot is being saved just prior to the end of a test execution it could happen the process was killed before the actual file was written to disk.
This fix addresses this issue by introducing a mutex with locks the instrumentation thread until it has been notified that the screenshot is saved
or it hits the 2 second timeout.